# perf: cache getElementShape() result